+++ engine/processReports/src/main/scala/pl/touk/nussknacker/processCounts/influxdb/ProcessBaseCounts.scala
@@ -2,10 +2,14 @@ package pl.touk.nussknacker.processCounts.influxdb
 
 case class ProcessBaseCounts(nodes: Map[String, Long]) {
 
+  //When using graphite protocol for reporting metrics
   //node ids in influx are different than original ones, i.e influx converts spaces and dots to dashes '-'
-  //that's wy we need reverse transformation
+  //that's why we need reverse transformation
+  //TODO: remove it when useLegacyMetrics flag is also removed
   def getCountForNodeId(nodeId: String) : Option[Long] = {
-    nodes.get(mapSpecialCharactersInfluxStyleNewVersions(nodeId))
+    nodes
+      .get(nodeId)
+      .orElse(nodes.get(mapSpecialCharactersInfluxStyleNewVersions(nodeId)))
       .orElse(nodes.get(mapSpecialCharactersInfluxStyleOldVersions(nodeId)))
   }